server/code/requirements.txt
flask
flask-sock
progress
prompt_toolkit
websocket-client